download page setup
